This significant improves the performance of subscriptions and unsubscriptions from listeners, at the expense of 5-10% slower event broadcasts.

```
benchmark                       old ns/op     new ns/op     delta
BenchmarkRecordAdd1K-8          4820          439           -90.89%
BenchmarkRecordAdd10K-8         164112        3901          -97.62%
BenchmarkRecordAdd100K-8        5522661       34703         -99.37%
BenchmarkRecordRemove1K-8       4882          78.8          -98.39%
BenchmarkRecordRemove10K-8      165169        78.3          -99.95%
BenchmarkRecordRemove100K-8     5663729       138           -100.00%
```

Previously, we had a list of subscriptions made copy-on-writes upon that list, while reads were atomic. However, this meant that subscriptions and unsubscriptions with many listeners became quite expensive.

In the new strategy, the list is instead a sequence of unsafe.Pointer's to Listener interfaces. We read these atomically when running broadcasts. The list can be 'sparse' in that there are nil pointers: when unsubscribing from an event, its pointer is nil'd in the list, or rewritten if the number of active listeners is below a density threshold value. When inserting new entries into the list, we'll try to fill an existing 'nil' slot if one exists, otherwise we'll grow the list (double it in size) which frees up lots of other slots. This means that amount of times we need to do a full copy-and-rewrite is in O(log n) for any sequence of subscriptions and unsubscriptions, rather than O(n), leading to performance improvements around 100-10000x with large lists

This includes a fuzzer for the new record list, which found no problems after running about 130 million sequences of insertions/deletions.