@connor4312 2018-06-15
Running this in prod for a week without issues. Merging it.